PROJQUAY-12005: fix(api): move superuser mirror health to superuser module
